### 2024-01-26 #623 @AAoritz
'nuts': Convert European Regional Data
Submitting Author Name: Moritz Hennicke
Submitting Author Github Handle: @AAoritz
Other Package Authors Github handles: @krausewe
Repository:  https://github.com/AAoritz/nuts/
Version submitted: 0.0.0.9000
Submission type: Standard

Version accepted: TBD

---

```
Package: nuts
Title: 'nuts': Convert European Regional Data
Version: 0.0.0.9000
Authors@R: c(person("Moritz", "Hennicke", email = "AAoritz@posteo.de", role = c("aut", "cre"),
	     person("Werner", "Krause", email = "werner.krause@uni-potsdam.de", role = c("aut"),
		   comment = c(ORCID = "0000-0002-5069-7964")))
Description: Motivated by changing administrative boundaries over time, the 'nuts' package can convert European regional data with NUTS codes between versions (2006, 2010, 2013, 2016 and 2021) and levels (NUTS 1, NUTS 2 and NUTS 3). The package uses spatial interpolation based on granular (100mx100m) population and land use data provided by the European Commission's Joint Research Center.
    distill,
    eurostat,
    formatR,
    ggalluvial,
    ggfittext,
    ggpubr,
    ggrepel,
    gridExtra,
    kableExtra,
    raster,
    RColorBrewer,
    sf,
    terra,
    tidyr
	crayon,
	stringr

```

Data munging: Linking regional data from different sources at the level of NUTS codes is often complicated by the usage of different versions or varying levels of geographical granularity across sources. The package can be used to harmonize NUTS versions and levels across different data sources using spatial interpolation.

Data validation and testing: The package includes routine tasks to test for the validity and completeness of NUTS codes.

Geospatial data: NUTS codes are the dominant format for European regional data.

The target audience are academics, journalists and data scientists interested in European regional data. Users who want to exploit changes within NUTS regions over time face the challenge that administrative boundaries are redrawn over time. The package enables the construction of consistent panel data across NUTS regions and over time through the harmonization of NUTS regions to one common version or level of granularity.

To our knowledge there is currently no package that is targeted at the conversion of NUTS versions using spatial interpolation.

The `regions` package allows for re-coding of NUTS codes from version to version without spatial interpolation. The package offers some code validation routines, but not an automated detection of the NUTS version used.

Yes
