+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   nuts-vignette: nuts-vignette.html
-last_built: 2024-01-26T17:12Z
+last_built: 2024-01-26T19:35Z
 urls:
   reference: https://aaoritz.github.io/nuts/reference
   article: https://aaoritz.github.io/nuts/articles
