### 2024-01-26 #623 @AAoritz
Hi rOpenSci Team and @ropensci-review-bot!

- We added a contributing file with this [commit](https://github.com/AAoritz/nuts/commit/d3c15290d28e1dd4e3d4e4a99c25d45bc831b8f6)
- The package uses an HTML vignette based on [`distill`](https://rstudio.github.io/distill/). The bot prefers in [`check-vignette.R`](https://github.com/ropensci-review-tools/pkgcheck/blob/e499f62755a76c237a50e26ec77d3bc695ace2a0/R/check-vignette.R#L58) `rmarkdown` HTML files. Should we switch to `rmarkdown`?

Thank your for your time and consideration!
