+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   nuts-vignette: nuts-vignette.html
-last_built: 2024-01-26T19:35Z
+last_built: 2024-01-29T09:49Z
 urls:
   reference: https://aaoritz.github.io/nuts/reference
   article: https://aaoritz.github.io/nuts/articles
