### 2024-01-29 #623 @jhollist
Yep, thanks @mpadge.  Will wait a bit of sending another check request to let those changes work through.

@AAoritz I think this looks good and is a fit for rOpenSci.  I will start working on finding a handling editor for your submission!
