### 2024-01-31 #623 @maelle
# Editor checks:

---

# Editor comments

Many thanks for your submission @AAoritz & @krausewe! Useful package! I know of the French equivalent to nuts, COGjugaison, so was vaguely aware of the idea https://antuki.github.io/COGugaison/ :smile_cat:

I have a few comments that I'd like to resolve or discuss before I proceed to the reviewer search.

## Contributing guide

- In the `CONTRIBUTING.md` file could you please remove the tidyverse specific bits such as " For more detailed info about contributing to this, and other tidyverse packages,"?

- I was surprised to see optional dependencies before Imports in `DESCRIPTION`. Would you mind running `desc::desc_normalize()` and see whether you like the result?

## Docs

- Instead of remotes in the README you could recommend pak, `pak::pak("AAoritz/nuts")` but in any case I don't think the `force = TRUE` argument is warranted, is it?

- Since there is only one vignette I'd recommend naming it "nuts.Rmd" so that it is automatically added in the pkgdown website navbar as "Get started". See https://pkgdown.r-lib.org/reference/build_articles.html#get-started

- Your images have alternative text but that text is not complete, it does not describe the image so that a screenreader user would miss a lot. https://www.w3.org/WAI/tutorials/images/informative/

- I was surprised by the short pipeline in https://github.com/AAoritz/nuts/blob/cf771ced02a20b226b56e6908959044ca62804e4/vignettes/nuts-vignette.Rmd#L247, could it avoid using the pipe, or if it does, could `filter(` go on a separate line? I feel it's an unusual pattern (but I might be wrong). I see this pattern of one-line pipelines in the tests too so it might be a personal preference! It's more noticeable/important in the user facing code but I won't impose _my_ personal style.

- The reference index is short but it might make sense to group the functions then the datasets. https://pkgdown.r-lib.org/reference/build_reference.html#reference-index

## User interface

- The function output of https://aaoritz.github.io/nuts/reference/classify_nuts.html#value is a list where you identify parts via their index. Could you add names to the output? Something like

```r
output <- list(
  data = data,
  versions_data = data_all_versions,
  missing_data = data_missing_nuts
)
```

so that then the user could use `output [["missing_data"]]` instead of `output[[3]]` which is less readable.

- Regarding the messages,
    - I see you use crayon for information but `stop()` for error messages. The crayon package has been superseeded by cli (see https://cli.r-lib.org/ and https://blog.r-hub.io/2023/11/30/cliff-notes-about-cli/), according to https://cran.r-project.org/web/packages/crayon/index.html. You could keep using crayon (it's still supported) but it might make sense switching, as cli provides really nice features. Code like

```r
stop("Input 'nuts_code' must be provided as a string.")
```

could become

```r
cli::cli_abort("Input {.arg nuts_code} must be provided as a string, not {.obj_type_friendly {nuts_code}}.")
```

(note that this function also requires importing rlang in DESCRIPTION)

and info messages could use `cli_alert_` functions https://cli.r-lib.org/reference/cli_alert.html

    - Most importantly, at the moment there is no way for the user to turn off messages. This is guidance we've just added to the development version of the dev guide so pretty new but I would still like this to be implemented as it might improve the quality of life of the users (including you, for instance in tests). See https://devdevguide.netlify.app/pkg_building#console-messages, https://blog.r-hub.io/2023/11/30/cliff-notes-about-cli/#how-to-make-cli-quiet-or-not if you use cli, and, hum, a draft blog post @mpadge and I are working on https://github.com/ropensci/roweb3/pull/717 Based on the interface tooling you end up selecting I am happy to provide tips about silencing if needed. Thanks for your understanding. :sweat_smile:

- Why does the package import whole packages rather than individual functions? Furthermore the imports should happen only once so they could happen in the package-level manual page source. You can create it with https://usethis.r-lib.org/reference/use_package_doc.html whose docs say "This .R file is also a good place for roxygen directives that apply to the whole package (vs. a specific function), such as global namespace tags like @importFrom."

- In lines such as https://github.com/AAoritz/nuts/blob/cf771ced02a20b226b56e6908959044ca62804e4/R/convert_nuts_version.R#L128 the code might benefit from the use of ["explaining variables"](https://blog.thepete.net/blog/2021/06/24/explaining-variable/), for instance (with a better name probably, again an example :sweat_smile: )

```r
not_enough_codes <- (length(data$from_code[check_nuts_codes]) < length(data$from_code) &&
               length(data$from_code[check_nuts_codes]) > 0)
if  (not_enough_codes)
```
(oh and the use of && instead of & see https://lintr.r-lib.org/reference/vector_logic_linter.html)

- You do not need `data()` to load package data in tests.

- I'd recommend not having top-level code in tests which means no code outside of `test_that(`. This makes tests more self contained. So in a helper file, `tests/testthat/helper-test-data.R` for instance you'd have the function

```r
manure_indic_DE_2003 <- function() {
manure %>%
  filter(nchar(geo) == 4) %>%
  filter(indic_ag == "I07A_EQ_Y") %>%
  select(-indic_ag) %>%
  filter(grepl("^DE", geo)) %>%
  filter(time == 2003) %>%
  select(-time)
}
```

and in `tests/testthat/test-classify_nuts.R` you'd have

```r
test_that("Needs geo var 1", {
  expect_error(
    manure_indic_DE_2003() %>% classify_nuts(nuts_code = NULL),
    "Input 'nuts_code' cannot be NULL."
  )
})

test_that("Needs geo var 2", {
  expect_error(manure_indic_DE_2003() %>% classify_nuts())
})

test_that("nuts_code not valid", {
  expect_error(
    manure_indic_DE_2003() %>% classify_nuts(nuts_code = 1),
    "Input 'nuts_code' must be provided as a string."
  )
})
```

See https://blog.r-hub.io/2020/11/18/testthat-utility-belt/ and https://r-pkgs.org/testing-design.html + https://r-pkgs.org/testing-advanced.html

- Instead of using `set.seed()`, in tests please use the withr version: https://withr.r-lib.org/reference/with_seed.html
---

Happy to answer any question and to discuss! Thanks again for submitting your package!
