### 2024-02-01 f5c4a5a
remove tidyverse specific bits in CONTRIBUTING.md
