### 2024-02-01 b80c8db
reorganize DESCRIPTION with desc::desc_normalize()
