+++ DESCRIPTION
@@ -1,17 +1,27 @@
 Package: nuts
 Title: 'nuts': Convert European Regional Data
 Version: 0.0.0.9000
-Authors@R: c(person("Moritz", "Hennicke", email = "AAoritz@posteo.de", role = c("aut", "cre"),
-		   comment = c(ORCID = "0000-0001-6811-1821")),
-	     person("Werner", "Krause", email = "werner.krause@uni-potsdam.de", role = c("aut"),
-		   comment = c(ORCID = "0000-0002-5069-7964")))
-Description: Motivated by changing administrative boundaries over time, the 'nuts' package can convert European regional data with NUTS codes between versions (2006, 2010, 2013, 2016 and 2021) and levels (NUTS 1, NUTS 2 and NUTS 3). The package uses spatial interpolation based on granular (100mx100m) population and land use data provided by the European Commission's Joint Research Center.  
+Authors@R: c(
+    person("Moritz", "Hennicke", , "AAoritz@posteo.de", role = c("aut", "cre"),
+           comment = c(ORCID = "0000-0001-6811-1821")),
+    person("Werner", "Krause", , "werner.krause@uni-potsdam.de", role = "aut",
+           comment = c(ORCID = "0000-0002-5069-7964"))
+  )
+Description: Motivated by changing administrative boundaries over time,
+    the 'nuts' package can convert European regional data with NUTS codes
+    between versions (2006, 2010, 2013, 2016 and 2021) and levels (NUTS 1,
+    NUTS 2 and NUTS 3). The package uses spatial interpolation based on
+    granular (100mx100m) population and land use data provided by the
+    European Commission's Joint Research Center.
+License: MIT + file LICENSE
 URL: https://AAoritz.github.io/nuts/
 BugReports: https://github.com/AAoritz/nuts/
-License: MIT + file LICENSE
-Encoding: UTF-8
-Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.1
+Depends: 
+    R (>= 3.5.0)
+Imports:
+    crayon,
+    dplyr,
+    stringr
 Suggests: 
     distill,
     eurostat,
@@ -32,12 +42,10 @@ Suggests:
     terra,
     testthat,
     tidyr
+VignetteBuilder: 
+    knitr
 Config/testthat/edition: 3
-VignetteBuilder: knitr
-Imports:
-	crayon,
-	dplyr,
-	stringr
-Depends: 
-    R (>= 3.5.0)
+Encoding: UTF-8
 LazyData: true
+Roxygen: list(markdown = TRUE)
+RoxygenNote: 7.3.1
