### 2024-02-01 51c2e44
README: install with pak
