+++ pkgdown.yml
@@ -2,8 +2,8 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
-  nuts-vignette: nuts-vignette.html
-last_built: 2024-01-29T09:49Z
+  nuts: nuts.html
+last_built: 2024-02-01T10:29Z
 urls:
   reference: https://aaoritz.github.io/nuts/reference
   article: https://aaoritz.github.io/nuts/articles
