### 2024-02-01 51f1cb3
Vignette: Add alt text
