### 2024-02-01 baf01ff
vignette: avoid short pipeline
