+++ vignettes/nuts.Rmd
@@ -121,10 +121,13 @@ class <- manure_indic %>%
   distinct(geo, .keep_all = T) %>%
   classify_nuts(data = ., nuts_code = "geo")
 
-transf <- class %>% convert_nuts_version( to_version = "2010" , variables = c( 'values' = 'absolute' ) , weight = 'artif_surf12' )
+transf <- class %>% 
+  convert_nuts_version( to_version = "2010" , variables = c( 'values' = 'absolute' ) , weight = 'artif_surf12' )
 
-small <- manure_indic %>% filter( geo %in% c( 'DED1' , 'DED3' ) )
-small <- small %>% mutate( artif_surf12 = case_when( geo == "DED1" ~ 98648
+small <- manure_indic %>% 
+  filter( geo %in% c( 'DED1' , 'DED3' ) )
+small <- small %>% 
+  mutate( artif_surf12 = case_when( geo == "DED1" ~ 98648
                                              , geo == "DED3" ~ 66786 + 5574
 ))
 
@@ -138,7 +141,9 @@ shape_06_n2 <- read_sf("NUTS_RG_20M_2006_3857.shp/NUTS_RG_20M_2006_3857.shp") %>
   full_join( manure_indic , by = c("NUTS_ID" = "geo")) %>%
   filter( str_detect( NUTS_ID , '^DED1|^DED3' ))
 
-shape_06_n2_centr <- shape_06_n2 %>% st_centroid( )  %>% left_join( small , by = c("NUTS_ID" = "geo"))
+shape_06_n2_centr <- shape_06_n2 %>% 
+  st_centroid( )  %>% 
+  left_join( small , by = c("NUTS_ID" = "geo"))
 
 p_initial = ggplot() +
   geom_sf(data = shape_06_n2 ,  linewidth = 1 , aes( fill = NUTS_ID )) +
@@ -156,7 +161,10 @@ p_initial = ggplot() +
         , axis.title = element_blank( )
   )
 
-DED33_centr <- shape_06_n3 %>% filter( NUTS_ID == "DED33" )  %>% st_centroid( ) %>% mutate( artif_surf12 = 5574 )
+DED33_centr <- shape_06_n3 %>% 
+  filter( NUTS_ID == "DED33" )  %>% 
+  st_centroid( ) %>% 
+  mutate( artif_surf12 = 5574 )
 
 shape_06_n2_step <- shape_06_n2 %>%
   mutate( artif_surf12 = case_when( NUTS_ID == 'DED3' ~ 66786
@@ -244,7 +252,8 @@ library(nuts)
 # Loading and subsetting Eurostat data
 data(patents, package = "nuts")
 
-pat.n2 <- patents %>% filter(nchar(geo) == 4) # NUTS-2 values
+pat.n2 <- patents %>% 
+  filter(nchar(geo) == 4) # NUTS-2 values
 
 pat.n2.mhab.12.no <- pat.n2 %>%
   filter(unit == "P_MHAB") %>% # Patents per one million inhabitants
@@ -322,7 +331,7 @@ gg_2006 = ggplot() +
     linewidth = .5
   ) +
   geom_sf(
-    data = no_2006 %>% filter(NUTS_ID %in% no_changes),
+    data = filter(no_2006, NUTS_ID %in% no_changes),
     color = 'red' ,
     fill = "#00000000"
   ) +
@@ -346,7 +355,7 @@ gg_2021 = ggplot() +
     linewidth = .5
   ) +
   geom_sf(
-    data = no_2021 %>% filter(NUTS_ID %in% no_changes),
+    data = filter(no_2021, NUTS_ID %in% no_changes),
     color = 'red' ,
     fill = "#00000000"
   ) +
[…]
