### 2024-02-02 a5bbea9
Change names of classify_nuts() output
