+++ tests/testthat/test-convert_nuts_version.R
@@ -40,7 +40,7 @@ test_that("data input not valid", {
       to_version = "2021",
       variables = c("values" = "absolute")
     ),
-    "Input 'data' must be a nuts.classified-object."
+    "Input `data` must be a nuts.classified-object."
   )
 })
 
@@ -62,7 +62,7 @@ test_that("variable not found", {
         to_version = "2021",
         variables = c("valuess" = "absolute")
       ),
-    "Input 'variables' not found in data."
+    "Input `variables` not found in the provided data frame."
   )
 })
 
@@ -88,7 +88,7 @@ test_that("to_version invalid", {
         to_version = "2020",
         variables = c("values" = "absolute")
       ),
-    "Input 'to_version' invalid. Make sure it is a string and one of the version years 2006, 2010, 2013, 2016 or 2021."
+    "Input `to_version` invalid. Make sure it is a string and one of the version years 2006, 2010, 2013, 2016 or 2021."
   )
 })
 
@@ -102,7 +102,7 @@ test_that("weight invalid", {
         variables = c("values" = "absolute"),
         weight = "pop19"
       ),
-    "Input 'weight' invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+    "Input `weight` invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
   )
 })
 
@@ -281,3 +281,4 @@ test_that("Feeding multiple NUTS versions within groups. Option most frequent.",
               c(1005, 5)
             )
           })
+
