+++ tests/testthat/test-classify_nuts.R
@@ -22,13 +22,13 @@ patents_nr_n2_2012 <- patents %>%
 # Run error tests
 test_that("data not valid", {
   expect_error(classify_nuts(data = 1 , nuts_code = "geo"),
-               "Input 'data' must be a data frame or tibble.")
+               "Input `data` must be a data frame or tibble, not a number.")
 })
 
 test_that("Needs geo var 1", {
   expect_error(
     manure_indic_DE_2003 %>% classify_nuts(nuts_code = NULL),
-    "Input 'nuts_code' cannot be NULL."
+    "Input `nuts_code` must be provided as a string, not NULL."
   )
 })
 
@@ -39,21 +39,21 @@ test_that("Needs geo var 2", {
 test_that("nuts_code not valid", {
   expect_error(
     manure_indic_DE_2003 %>% classify_nuts(nuts_code = 1),
-    "Input 'nuts_code' must be provided as a string."
+    "Input `nuts_code` must be provided as a string, not a number."
   )
 })
 
 test_that("nuts_code name not found", {
   expect_error(
     manure_indic_DE_2003 %>% classify_nuts(nuts_code = "geoo"),
-    "Input 'nuts_code' not found in the provided data frame."
+    "Input `nuts_code` not found in the provided data frame."
   )
 })
 
 test_that("ties not valid", {
   expect_error(
     manure_indic_DE_2003 %>% classify_nuts(nuts_code = "geo", ties = "most_recentt"),
-    "Input 'ties' must be 'most_recent' or 'oldest'."
+    "Input `ties` must be 'most_recent' or 'oldest'."
   )
 })
 
@@ -64,7 +64,7 @@ test_that("NUTS codes not valid", {
         mutate(geo = str_remove(geo, "[A-Z]")) %>%
         classify_nuts(nuts_code = "geo")
     ) %>%
-      grepl("Variable with NUTS codes contains invalid codes", .),
+      grepl("Variable `geo` contains invalid NUTS codes.", .),
     TRUE
   )
 })
@@ -87,7 +87,7 @@ test_that("grouping variable not found", {
   expect_error(
     manure_indic_DE_2003 %>%
       classify_nuts(nuts_code = "geo", group_vars = "group"),
-    "Input 'group_vars' not found in the provided data frame."
+    "Input `group_vars` not found in the provided data frame."
   )
 })
 
@@ -201,3 +201,4 @@ test_that("One missing NUTS code", {
     c(1, 4)
   )
 })
+
