### 2024-02-02 434280b
Call cli in vignette
