+++ vignettes/nuts.Rmd
@@ -106,6 +106,7 @@ library(ggplot2)
 library(ggpubr)
 library(dplyr)
 library(stringr)
+library(cli)
 library(formatR)
 
 data(manure, package = "nuts")
