### 2024-02-02 5ba93a4
Import cli in functions
