### 2024-02-03 e2ba58f
move top-level codes in tests as functions to helper file
