+++ tests/testthat/test-classify_nuts.R
@@ -1,58 +1,44 @@
-# Load testing data
-data("manure")
-data("patents")
-
-manure_indic_DE_2003 <- manure %>%
-  filter(nchar(geo) == 4) %>%
-  filter(indic_ag == "I07A_EQ_Y") %>%
-  select(-indic_ag) %>%
-  filter(grepl("^DE", geo)) %>%
-  filter(time == 2003) %>%
-  select(-time)
-
-manure_indic <- manure %>%
-  filter(nchar(geo) == 4) %>%
-  filter(indic_ag == "I07A_EQ_Y") %>%
-  select(-indic_ag)
-
-patents_nr_n2_2012 <- patents %>%
-  filter(unit == "NR", nchar(geo) == 4, time == 2012)
-
-
 # Run error tests
 test_that("data not valid", {
-  expect_error(classify_nuts(data = 1 , nuts_code = "geo"),
-               "Input `data` must be a data frame or tibble, not a number.")
+  expect_error(
+    classify_nuts(data = 1 , nuts_code = "geo"),
+    "Input `data` must be a data frame or tibble, not a number."
+  )
 })
 
 test_that("Needs geo var 1", {
   expect_error(
-    manure_indic_DE_2003 %>% classify_nuts(nuts_code = NULL),
+    manure_indic_2_DE_2003() %>%
+      classify_nuts(nuts_code = NULL),
     "Input `nuts_code` must be provided as a string, not NULL."
   )
 })
 
 test_that("Needs geo var 2", {
-  expect_error(manure_indic_DE_2003 %>% classify_nuts())
+  expect_error(manure_indic_2_DE_2003() %>%
+                 classify_nuts())
 })
 
 test_that("nuts_code not valid", {
   expect_error(
-    manure_indic_DE_2003 %>% classify_nuts(nuts_code = 1),
+    manure_indic_2_DE_2003() %>%
+      classify_nuts(nuts_code = 1),
     "Input `nuts_code` must be provided as a string, not a number."
   )
 })
 
 test_that("nuts_code name not found", {
   expect_error(
-    manure_indic_DE_2003 %>% classify_nuts(nuts_code = "geoo"),
+    manure_indic_2_DE_2003() %>%
+      classify_nuts(nuts_code = "geoo"),
     "Input `nuts_code` not found in the provided data frame."
   )
 })
 
 test_that("ties not valid", {
   expect_error(
-    manure_indic_DE_2003 %>% classify_nuts(nuts_code = "geo", ties = "most_recentt"),
+    manure_indic_2_DE_2003() %>%
+      classify_nuts(nuts_code = "geo", ties = "most_recentt"),
     "Input `ties` must be 'most_recent' or 'oldest'."
   )
 })
@@ -60,7 +46,7 @@ test_that("ties not valid", {
 test_that("NUTS codes not valid", {
   expect_equal(
     expect_error(
-      manure_indic_DE_2003 %>%
+      manure_indic_2_DE_2003() %>%
         mutate(geo = str_remove(geo, "[A-Z]")) %>%
         classify_nuts(nuts_code = "geo")
     ) %>%
@@ -85,37 +71,44 @@ test_that("Multiple levels", {
 
 test_that("grouping variable not found", {
   expect_error(
-    manure_indic_DE_2003 %>%
+    manure_indic_2_DE_2003() %>%
       classify_nuts(nuts_code = "geo", group_vars = "group"),
     "Input `group_vars` not found in the provided data frame."
   )
 })
 
 test_that("no grouping variable was used or NUTS codes are not unique", {
-  expect_equal(expect_error(classify_nuts(data = manure_indic, nuts_code = "geo")) %>%
[…]

[…]
