+++ R/classify_nuts.R
@@ -16,10 +16,6 @@
 #'
 #' The output can be passed to [convert_nuts_version()] to convert data across NUTS versions and [convert_nuts_level()] to aggregate across NUTS levels.
 #'
-#' @import cli
-#' @import rlang
-#' @import dplyr
-#' @import stringr
 #'
 #' @examples
 #' library(dplyr)
