+++ R/convert_nuts_level.R
@@ -13,10 +13,6 @@
 #'
 #' @return A tibble containing NUTS codes, aggregated variable values, and possibly grouping variables.
 #'
-#' @import cli
-#' @import rlang
-#' @import dplyr
-#' @import stringr
 #'
 #' @examples
 #' library(dplyr)
