+++ R/nuts-package.R
@@ -1,0 +1,10 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @import cli
+#' @import dplyr
+#' @import rlang
+#' @import stringr
+## usethis namespace: end
+NULL
