+++ R/convert_nuts_version.R
@@ -13,10 +13,6 @@
 #'
 #' @return A tibble containing NUTS codes, converted variable values, and possibly grouping variables.
 #'
-#' @import cli
-#' @import rlang
-#' @import dplyr
-#' @import stringr
 #'
 #' @examples
 #' library(dplyr)
