+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(classify_nuts)
 export(convert_nuts_level)
 export(convert_nuts_version)
-import(crayon)
+import(cli)
 import(dplyr)
+import(rlang)
 import(stringr)
