### 2024-02-03 fd80dd1
compile nuts-package.Rd
