### 2024-02-03 3156569
use withr::with_preserve_seed in tests, suggest withr
