### 2024-02-03 0d9e3ad
use with_seed instead of with_preserve_seed in testthat helper
