### 2024-02-03 4bc0303
add explaining variables, double &&, format code blocks
