+++ R/convert_nuts_level.R
@@ -46,7 +46,9 @@ convert_nuts_level <-
     #------------------------
     # Input checks
     if (any(class(data) != "nuts.classified"))
-      cli_abort("Input {.arg data} must be a nuts.classified-object, not {.obj_type_friendly {data}}.")
+      cli_abort(
+        "Input {.arg data} must be a nuts.classified-object, not {.obj_type_friendly {data}}."
+      )
 
     if (is.null(variables))
       cli_abort("Input {.arg variables} cannot be NULL.")
@@ -57,17 +59,21 @@ convert_nuts_level <-
     if (any(!(unlist(variables) %in% c("absolute", "relative"))))
       cli_abort("Variable type(s) not found. Use one of the following: 'absolute' or 'relative'.")
 
-    if (!is.numeric(to_level)|!to_level %in% c(1, 2))
+    if (!is.numeric(to_level) | !to_level %in% c(1, 2))
       cli_abort("Input {.arg to_level} invalid. Must be 1 or 2.")
 
     if (!is.null(weight)) {
       if (!weight %in% c("areaKm", "pop18", "pop11", "artif_surf18",
                          "artif_surf12"))
-        cli_abort("Input {.arg weight} invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'.")
+        cli_abort(
+          "Input {.arg weight} invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+        )
     }
 
     if (!is.logical(missing_rm))
-      cli_abort("Input {.arg missing_rm} invalid. Must be TRUE/FALSE, not {.obj_type_friendly {missing_rm}}.")
+      cli_abort(
+        "Input {.arg missing_rm} invalid. Must be TRUE/FALSE, not {.obj_type_friendly {missing_rm}}."
+      )
 
     if (!multiple_versions %in% c("break", "most_frequent"))
       cli_abort("Input {.arg multiple_versions} invalid. Must be 'break' or 'most_frequent'.")
@@ -81,7 +87,9 @@ convert_nuts_level <-
     # Check whether user is trying to aggregate to a lower level than
     # current one
     if (data$from_level[1] < to_level)
-      cli_abort("Provided NUTS codes are on a higher level than input {.arg to_level}. Only data aggregation to a higher NUTS level allowed.")
+      cli_abort(
+        "Provided NUTS codes are on a higher level than input {.arg to_level}. Only data aggregation to a higher NUTS level allowed."
+      )
 
     if (data$from_level[1] == to_level)
       cli_abort("NUTS codes already at level {to_level}.")
@@ -89,7 +97,9 @@ convert_nuts_level <-
     # Check if NUTS codes can be converted
     all_nuts_codes <- get("all_nuts_codes")
     check_nuts_codes <- data$from_code %in% all_nuts_codes$code
-    if (length(data$from_code[check_nuts_codes]) == 0)
+    nr_nuts_codes_recognized <-
+      length(data$from_code[check_nuts_codes])
+    if (nr_nuts_codes_recognized == 0)
       cli_abort("NUTS codes are not recognized and cannot be converted.")
 
     # CONVERSION POSSIBLE
@@ -110,14 +120,19 @@ convert_nuts_level <-
     ))
 
     # Check which NUTS codes can be converted
-    if (length(data$from_code[check_nuts_codes]) == length(data$from_code)) {
+    nr_nuts_codes_recognized <-
+      length(data$from_code[check_nuts_codes])
+    nr_nuts_codes <- length(data$from_code)
+    if (nr_nuts_codes_recognized == nr_nuts_codes) {
[…]
