+++ R/convert_nuts_version.R
@@ -61,7 +61,9 @@ convert_nuts_version <-
     #------------------------
     # Input checks
     if (!any(class(data) == "nuts.classified"))
-      cli_abort("Input {.arg data} must be a nuts.classified-object, not {.obj_type_friendly {data}}.")
+      cli_abort(
+        "Input {.arg data} must be a nuts.classified-object, not {.obj_type_friendly {data}}."
+      )
 
     if (is.null(variables))
       cli_abort("Input {.arg variables} cannot be NULL.")
@@ -73,16 +75,22 @@ convert_nuts_version <-
       cli_abort("Variable type(s) not found. Use one of the following: 'absolute' or 'relative'.")
 
     if (!to_version %in% c("2006", "2010", "2013", "2016", "2021"))
-      cli_abort("Input {.arg to_version} invalid. Make sure it is a string and one of the version years 2006, 2010, 2013, 2016 or 2021.")
+      cli_abort(
+        "Input {.arg to_version} invalid. Make sure it is a string and one of the version years 2006, 2010, 2013, 2016 or 2021."
+      )
 
     if (!is.null(weight)) {
       if (!weight %in% c("areaKm", "pop18", "pop11", "artif_surf18",
                          "artif_surf12"))
-        cli_abort("Input {.arg weight} invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'.")
+        cli_abort(
+          "Input {.arg weight} invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+        )
     }
 
     if (!is.logical(missing_rm))
-      cli_abort("Input {.arg missing_rm} invalid. Must be TRUE/FALSE, not {.obj_type_friendly {missing_rm}}.")
+      cli_abort(
+        "Input {.arg missing_rm} invalid. Must be TRUE/FALSE, not {.obj_type_friendly {missing_rm}}."
+      )
 
     if (!multiple_versions %in% c("break", "most_frequent"))
       cli_abort("Input {.arg multiple_versions} invalid. Must be 'break' or 'most_frequent'.")
@@ -95,8 +103,10 @@ convert_nuts_version <-
     # Check if NUTS codes can be converted
     all_nuts_codes <- get("all_nuts_codes")
     check_nuts_codes <- data$from_code %in% all_nuts_codes$code
-    if (length(data$from_code[check_nuts_codes]) == 0)
-      cli_abort("\nNUTS codes are not recognized and cannot be converted.")
+    nr_nuts_codes_recognized <-
+      length(data$from_code[check_nuts_codes])
+    if (nr_nuts_codes_recognized == 0)
+      cli_abort("NUTS codes are not recognized and cannot be converted.")
 
     # CONVERSION POSSIBLE
     #----------------------
@@ -106,29 +116,37 @@ convert_nuts_version <-
     # CONVERSION BETWEEN DIFFERENT NUTS VERSIONS
     cat(col_blue(
       paste0(
-        "\n=> Converting NUTS codes in version(s) " , col_red(paste0(sort(
+        "\n=> Converting NUTS codes in version(s) " ,
+        col_red(paste0(sort(
           unique(data$from_version[!is.na(data$from_version)])
-        ), collapse = ", ")) , " to version ",
+        ), collapse = ", ")) ,
+        " to version ",
         col_red(to_version),
         "."
       )
     ))
 
     # Check which NUTS codes can be converted
-    if (length(data$from_code[check_nuts_codes]) == length(data$from_code)) {
+    nr_nuts_codes_recognized <-
+      length(data$from_code[check_nuts_codes])
[…]
