### 2024-02-03 f11a080
add reference index to pkgdown grouping functions and data
