+++ _pkgdown.yml
@@ -3,3 +3,12 @@ template:
   bootstrap: 5
 home:
   title: Convert European Regional Data in R
+reference:
+- title: Functions
+  desc:  Functions that classify and convert regional data
+- contents:
+    starts_with("convert_|classify_")
+- title: Data
+  desc:  Data sets required for conversion and examples
+- contents:
+    starts_with("all_nuts_|cross_|manure|patents")
