+++ R/nuts-package.R
@@ -3,8 +3,45 @@
 
 ## usethis namespace: start
 #' @import cli
-#' @import dplyr
+#' @import magrittr
 #' @import rlang
-#' @import stringr
+#' @import tibble
+#' @importFrom dplyr anti_join
+#' @importFrom dplyr arrange
+#' @importFrom dplyr cross_join
+#' @importFrom dplyr distinct
+#' @importFrom dplyr filter
+#' @importFrom dplyr full_join
+#' @importFrom dplyr group_by
+#' @importFrom dplyr group_by_at
+#' @importFrom dplyr inner_join
+#' @importFrom dplyr left_join
+#' @importFrom dplyr right_join
+#' @importFrom dplyr mutate
+#' @importFrom dplyr mutate_at
+#' @importFrom dplyr pull
+#' @importFrom dplyr relocate
+#' @importFrom dplyr rename
+#' @importFrom dplyr select
+#' @importFrom dplyr slice
+#' @importFrom dplyr summarize
+#' @importFrom dplyr summarise
+#' @importFrom dplyr summarize_at
+#' @importFrom dplyr summarise_at
+#' @importFrom dplyr ungroup
+#' @importFrom dplyr bind_rows
+#' @importFrom dplyr n_distinct
+#' @importFrom dplyr vars
+#' @importFrom dplyr any_of
+#' @importFrom dplyr all_of
+#' @importFrom dplyr if_any
+#' @importFrom dplyr n
+#' @importFrom dplyr case_when
+#' @importFrom dplyr join_by
+#' @importFrom dplyr desc
+#' @importFrom dplyr arrange_at
+#' @importFrom glue glue
+#' @importFrom lifecycle deprecated
+#' @importFrom stringr str_detect
 ## usethis namespace: end
 NULL
