+++ NAMESPACE
@@ -4,6 +4,41 @@ export(classify_nuts)
 export(convert_nuts_level)
 export(convert_nuts_version)
 import(cli)
-import(dplyr)
+import(magrittr)
 import(rlang)
-import(stringr)
+import(tibble)
+importFrom(dplyr,all_of)
+importFrom(dplyr,anti_join)
+importFrom(dplyr,any_of)
+importFrom(dplyr,arrange)
+importFrom(dplyr,arrange_at)
+importFrom(dplyr,bind_rows)
+importFrom(dplyr,case_when)
+importFrom(dplyr,cross_join)
+importFrom(dplyr,desc)
+importFrom(dplyr,distinct)
+importFrom(dplyr,filter)
+importFrom(dplyr,full_join)
+importFrom(dplyr,group_by)
+importFrom(dplyr,group_by_at)
+importFrom(dplyr,if_any)
+importFrom(dplyr,inner_join)
+importFrom(dplyr,join_by)
+importFrom(dplyr,left_join)
+importFrom(dplyr,mutate)
+importFrom(dplyr,mutate_at)
+importFrom(dplyr,n)
+importFrom(dplyr,n_distinct)
+importFrom(dplyr,pull)
+importFrom(dplyr,relocate)
+importFrom(dplyr,rename)
+importFrom(dplyr,right_join)
+importFrom(dplyr,select)
+importFrom(dplyr,slice)
+importFrom(dplyr,summarize)
+importFrom(dplyr,summarize_at)
+importFrom(dplyr,ungroup)
+importFrom(dplyr,vars)
+importFrom(glue,glue)
+importFrom(lifecycle,deprecated)
+importFrom(stringr,str_detect)
