### 2024-02-06 bfdb7c5
Change cli_warn() to cli_alert_warning()
