+++ R/convert_nuts_level.R
@@ -113,8 +113,8 @@ convert_nuts_level <-
     # CONVERSION POSSIBLE
     #----------------------
     # Welcome information
-    cli_alert_info(c("{.blue Converting level of NUTS codes}",
-                     "\n{.blue --------------------------------}"))
+    cli_text("{.blue Converting level of NUTS codes}")
+    cli_text("{.blue ------------------------------}")
     # CONVERSION BETWEEN DIFFERENT NUTS LEVELS
     cli_alert_info("{.blue => Aggregate from NUTS regional level {.red {data$from_level[1]}} to {.red {to_level}}}.")
 
@@ -126,7 +126,7 @@ convert_nuts_level <-
       cli_alert_info("{.blue => All NUTS codes can be converted.}")
     } else if (nr_nuts_codes_recognized < nr_nuts_codes &&
                nr_nuts_codes_recognized > 0) {
-      cli_warn("{.blue => These NUTS codes cannot be converted and {.red are dropped} from the dataaset: {.red {unique(data$from_code[!check_nuts_codes])}}.}")
+      cli_alert_warning("{.blue => These NUTS codes cannot be converted and {.red are dropped} from the dataset: {.red {unique(data$from_code[!check_nuts_codes])}}.}")
       data <- data[check_nuts_codes, ]
     }
 
@@ -144,7 +144,7 @@ convert_nuts_level <-
     # Use data_versions which is sorted for most frequent version within group
     if (multi_versions_A > multi_versions_B) {
       cli_alert_info("{.blue => Within {.red groups} defined by {.red {group_vars}}}")
-      cli_warn("{.blue ==> {.red Multiple} NUTS code versions.}")
+      cli_alert_warning("{.blue ==> {.red Multiple} NUTS code versions.}")
 
       if (multiple_versions == "break") {
         cli_abort(
@@ -219,8 +219,8 @@ convert_nuts_level <-
     }
 
     if (nrow(missing) > 0) {
-      cli_warn(c("{.blue => {.red Missing} NUTS codes in data.}",
-                 "{.blue ==> No values are calculated for regions associated with missing NUTS codes. Ensure that the input data is complete.}"))
+      cli_alert_warning("{.blue => {.red Missing} NUTS codes in data.}")
+      cli_alert_warning("{.blue ==> No values are calculated for regions associated with missing NUTS codes. Ensure that the input data is complete.}")
     } else if (nrow(missing) == 0) {
       cli_alert_info("{.blue => No missing NUTS codes.}")
     }
