### 2024-02-06 9a93ff1
Option to silence cli
