+++ man/classify_nuts.Rd
@@ -8,7 +8,8 @@ classify_nuts(
   data = data,
   nuts_code = nuts_code,
   group_vars = NULL,
-  ties = "most_recent"
+  ties = "most_recent",
+  quiet = FALSE
 )
 }
 \arguments{
@@ -22,6 +23,8 @@ duplicate NUTS codes within groups allowed.}
 
 \item{ties}{Picks \code{'most_recent'} or \code{'oldest'} version when overlap is identical across multiple NUTS versions. \code{'most_recent'}
 by default.}
+
+\item{quiet}{Suppress messages and warnings. \code{TRUE} by default.}
 }
 \value{
 A list of three tibbles. The first tibble contains the original data with the classified NUTS version, level, and country.
