+++ NAMESPACE
@@ -12,6 +12,7 @@ importFrom(dplyr,anti_join)
 importFrom(dplyr,any_of)
 importFrom(dplyr,arrange)
 importFrom(dplyr,arrange_at)
+importFrom(dplyr,bind_cols)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,case_when)
 importFrom(dplyr,cross_join)
@@ -35,6 +36,8 @@ importFrom(dplyr,rename)
 importFrom(dplyr,right_join)
 importFrom(dplyr,select)
 importFrom(dplyr,slice)
+importFrom(dplyr,summarise)
+importFrom(dplyr,summarise_at)
 importFrom(dplyr,summarize)
 importFrom(dplyr,summarize_at)
 importFrom(dplyr,ungroup)
