+++ R/nuts-package.R
@@ -44,5 +44,6 @@
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 #' @importFrom stringr str_detect
+#' @importFrom stringr str_remove
 ## usethis namespace: end
 NULL
