+++ NAMESPACE
@@ -45,3 +45,4 @@ importFrom(dplyr,vars)
 importFrom(glue,glue)
 importFrom(lifecycle,deprecated)
 importFrom(stringr,str_detect)
+importFrom(stringr,str_remove)
