### 2024-02-07 4bcb540
adapt tests to modified cli error messages
