+++ NAMESPACE
@@ -4,7 +4,6 @@ export(classify_nuts)
 export(convert_nuts_level)
 export(convert_nuts_version)
 import(cli)
-import(magrittr)
 import(rlang)
 import(tibble)
 importFrom(dplyr,all_of)
@@ -44,5 +43,6 @@ importFrom(dplyr,ungroup)
 importFrom(dplyr,vars)
 importFrom(glue,glue)
 importFrom(lifecycle,deprecated)
+importFrom(magrittr,"%>%")
 importFrom(stringr,str_detect)
 importFrom(stringr,str_remove)
