+++ DESCRIPTION
@@ -26,8 +26,7 @@ Imports:
     magrittr,
     rlang,
     stringr,
-    tibble,
-    withr
+    tibble
 Suggests: 
     distill,
     eurostat,
@@ -47,7 +46,8 @@ Suggests:
     sf,
     terra,
     testthat,
-    tidyr
+    tidyr,
+    withr
 VignetteBuilder: 
     knitr
 Config/testthat/edition: 3
