+++ R/convert_nuts_level.R
@@ -10,7 +10,7 @@
 #' @param missing_rm Boolean that is FALSE by default. TRUE removes regional flows that depart from missing NUTS codes.
 #' @param multiple_versions By default equal to `'break'`, throwing an error when providing multiple NUTS versions within groups.
 #' If set to `'most_frequent'` data is converted using the best-matching NUTS version.
-#' @param quiet Suppress messages and warnings. `TRUE` by default.
+#' @param quiet Suppress messages and warnings. `FALSE` by default.
 #'
 #' @return A tibble containing NUTS codes, aggregated variable values, and possibly grouping variables.
 #'
