### 2024-02-07 c21b26b
group and restructure cli messages in classify and convert version
