### 2024-02-07 7859e02
bundle cli messages in list in convert_nuts_level
