+++ tests/testthat/test-convert_nuts_level.R
@@ -280,4 +280,3 @@ test_that("Feeding multiple NUTS versions within groups. Option most frequent.",
               c(52, 4)
             )
           })
-
