+++ R/convert_nuts_level.R
@@ -119,21 +119,20 @@ convert_nuts_level <-
 
     # CONVERSION POSSIBLE
     #----------------------
-    # Welcome information
-    cli_text("{.blue Converting level of NUTS codes}")
-    cli_text("{.blue ------------------------------}")
     # CONVERSION BETWEEN DIFFERENT NUTS LEVELS
-    cli_alert_info("{.blue => Aggregate from NUTS regional level {.red {data$from_level[1]}} to {.red {to_level}}}.")
+    aggregate_from_level <- data$from_level[1]
+    message_conversion_levels <- c("i" = "{.blue Aggregate from NUTS regional level {.red {aggregate_from_level}} to {.red {to_level}}}.")
 
     # Check which NUTS codes can be converted
     nr_nuts_codes_recognized <-
       length(data$from_code[check_nuts_codes])
     nr_nuts_codes <- length(data$from_code)
+    dropped_codes <- unique(data$from_code[!check_nuts_codes])
     if (nr_nuts_codes_recognized == nr_nuts_codes) {
-      cli_alert_info("{.blue => All NUTS codes can be converted.}")
+      message_can_be_converted <- c("v" = "{.blue All NUTS codes can be converted.}")
     } else if (nr_nuts_codes_recognized < nr_nuts_codes &&
                nr_nuts_codes_recognized > 0) {
-      cli_alert_warning("{.blue => These NUTS codes cannot be converted and {.red are dropped} from the dataset: {.red {unique(data$from_code[!check_nuts_codes])}}.}")
+      message_can_be_converted <- c("x" = "{.blue These NUTS codes cannot be converted and {.red are dropped}: {.red {dropped_codes}}.}")
       data <- data[check_nuts_codes, ]
     }
 
@@ -149,37 +148,33 @@ convert_nuts_level <-
       nrow()
 
     # Use data_versions which is sorted for most frequent version within group
-    if (multi_versions_A > multi_versions_B) {
-      cli_alert_info("{.blue => Within {.red groups} defined by {.red {group_vars}}}")
-      cli_alert_warning("{.blue ==> {.red Multiple} NUTS code versions.}")
+    if (multi_versions_A > multi_versions_B && multiple_versions == "break") {
 
-      if (multiple_versions == "break") {
-        cli_abort(
-          c(
-            "Mixed NUTS versions within groups!",
-            "Please make sure the data contains only one version per group. Alternatively, keep only the codes belonging to the 'most_frequent' version using the argument {.arg multiple_versions}."
-          )
+      cli_abort(
+        c(
+          "Mixed NUTS versions within groups!"
+          ,
+          "Please make sure the data contains only one version per group. Alternatively, keep only the codes belonging to the 'most_frequent' version using the argument 'multiple_versions'."
         )
+      )
 
-      } else if (multiple_versions == "most_frequent") {
-        data_versions <- data_versions %>%
-          group_by_at(vars(any_of(c(group_vars)))) %>%
-          slice(1) %>%
-          ungroup()
+    } else if (multi_versions_A > multi_versions_B && multiple_versions == "most_frequent") {
+      data_versions <- data_versions %>%
+        group_by_at(vars(any_of(c(group_vars)))) %>%
+        slice(1) %>%
+        ungroup()
 
-        data_multi_versions <-
-          anti_join(data, data_versions, by = c("from_version", group_vars))
-        data <-
-          inner_join(data, data_versions, by = c("from_version", group_vars))
+      data_multi_versions <-
+        anti_join(data, data_versions, by = c("from_version", group_vars))
+      data <-
+        inner_join(data, data_versions, by = c("from_version", group_vars))
 
[…]
