+++ man/classify_nuts.Rd
@@ -24,7 +24,7 @@ duplicate NUTS codes within groups allowed.}
 \item{ties}{Picks \code{'most_recent'} or \code{'oldest'} version when overlap is identical across multiple NUTS versions. \code{'most_recent'}
 by default.}
 
-\item{quiet}{Suppress messages and warnings. \code{TRUE} by default.}
+\item{quiet}{Suppress messages and warnings. \code{FALSE} by default.}
 }
 \value{
 A list of three tibbles. The first tibble contains the original data with the classified NUTS version, level, and country.
