### 2024-02-08 #623 @krausewe
Hi @maelle!

We have addressed the very helpful remarks & recommendations.

Please find below the list of commits.

We are looking forward to the next steps!

# Contributing guide
- Remove tidyverse specific points in `CONTRIBUTING`: https://github.com/AAoritz/nuts/commit/f5c4a5a4049c52c9f34dd253d99a0eb2515af164
- Normalize `DESCRIPTION`: https://github.com/AAoritz/nuts/commit/b80c8db0ad253a631b6b06f8fa20a61523c0d218

# Docs
- Change install to `pak` in `README`: https://github.com/AAoritz/nuts/commit/51c2e44addb5b5052421b967cb9ddc11837fb6af
- Rename vignette to `nuts.Rmd`: https://github.com/AAoritz/nuts/commit/3fe96c87ece74a6e28d35f350302d282523238e8
- Include alt-text of images in vignette: https://github.com/AAoritz/nuts/commit/51f1cb3c67b1e85acaef84dc3937620c963a3985
- Add line breaks for one-line pipelines: https://github.com/AAoritz/nuts/commit/baf01ff9227695fa60e95cc6ee61c7d6c3189202
- Group functions and data `pkgdown`: https://github.com/AAoritz/nuts/commit/f11a08070d0c1356339936c742165cfe6a551c34

# User interface
- Simplify names of output list of `classify`: https://github.com/AAoritz/nuts/commit/a5bbea99ace46bd32be7bc0628d239ebe0f319f4
- Use `cli` messages: https://github.com/AAoritz/nuts/commit/8cb9be6e0fa526f1fc43cbaeb8ab5b3bec89c8e8
- Silencing messages: https://github.com/AAoritz/nuts/commit/9a93ff174389a922ca7bae39daa039d93eb7454b

# Code
- Import functions as opposed to entire packages: https://github.com/AAoritz/nuts/commit/d2744d97cd43dd7a45c52416640633cfd604d34f
- Use explaining variables and double `&&`: https://github.com/AAoritz/nuts/commit/4bc0303f8b2be824ac554d31cb772ac27d242140

# Tests
- Don't load `data()`: https://github.com/AAoritz/nuts/commit/e2ba58f62169b8c84b33817faf305dfc6aeb6c09
- Include helper file that prepares data: https://github.com/AAoritz/nuts/commit/e2ba58f62169b8c84b33817faf305dfc6aeb6c09
- Replace `set.seed()` with `withr`: https://github.com/AAoritz/nuts/commit/0d9e3ad4084780c69179f52afd21c30b1eedba6c
