### 2024-02-14 47ee441
implement verbosity control with local options
