+++ R/convert_nuts_level.R
@@ -266,15 +266,19 @@ convert_nuts_level <-
 
     # Console Message
     #-----------------
-    cli_h1("Converting version of NUTS codes")
-    cli_bullets(
-      c("{.blue Within {.red groups} defined by {.red {group_vars}}:}",
-        message_conversion_levels,
-        message_can_be_converted,
-        message_multiple_versions,
-        message_missing_codes
+    is_verbose_mode <- (getOption("nuts.verbose", "quiet") == "verbose")
+    if (is_verbose_mode) {
+      cli_h1("Converting version of NUTS codes")
+      cli_bullets(
+        c(
+          "{.blue Within {.red groups} defined by {.red {group_vars}}:}",
+          message_conversion_levels,
+          message_can_be_converted,
+          message_multiple_versions,
+          message_missing_codes
+        )
       )
-    )
+    }
 
     return(data)
   }
