### 2024-02-14 6438c6a
remove quiet option from function arguments
