### 2024-02-14 #623 @mpadge
@AAoritz Would you please be so kind as to ask that question on [our discussion forum](https://discuss.ropensci.org/t/ropensci-please-shut-up-verbosity-control-in-packages/3790/12)? The issue of where to document verbosity control is important, and would be good to have in that general context, rather than here.
