### 2024-02-18 #623 @jospueyo
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

-   [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

The package is useful, very useful indeed, and also well designed. Most of my comments below are just suggestions that IMHO will make the package more intuitive, more robust, or more easily maintainable. Don't take them as mandatory. At the end, this is your package and you should have the last word.

Disclaimer: Apologies if some comments sound though or rude. English is not my native language.

Function naming does not follow rOpenSci package guidelines. Specifically, the design `object_verb()`. Please, start your exported functions with `nuts_`, this also facilitates autocompletion.

Variables in vignettes should use underscores and not points, that can clash with S3 methods.

`group_vars` argument could be changed to `.by`, to be consinstent with `dplyr` functions.

`convert_nuts_levels` might be changed to `nuts_aggregate` to be explicit about that the level change is only aggregation and not disaggregation.

Provide wrapper functions to inspect sub-elements of `classifiy_nuts`, something like:

- `nuts_get_data()`
- `nuts_get_version()`
- `nuts_get_missing()`

`missing_rm` could also provide some imputation technique, such as using the mean. In the future, it might be linked to imputation packages such as `mice`, for instance.

There are some inconsistencies in your constructions that can make the package harder to mantain, especially if there is a maintainer change in the future. For instance, sometimes you are explicit inside your conditions: `if(condition == FALSE) do something...` vs `if(!condition) do something...`. You should check the constructions you used and try to be consistent. BTW, you should avoid using `T` and `F`, replace them with `TRUE` and `FALSE`.

You should not supply default values in mandatory arguments. Otherwise, if you run the function without providing these values, it raises a cryptographic error instead of the usual `Error in classify_nuts() : argument "data" is missing, with no default`.

When an argument accepts a set of values, such as`multiple_versions` or `weights`, it’s preferable to pass a vector with all options as default to show all possible values. E.g. `multiple_versions = c('break', 'most_frequent')`. Then, you use the first element, which would be `break` by default. BTW, `break` is a concept that not anybody might be familiar with. Maybe, using `error` would be more intuitive.

When you check if an argument is one of the accepted values, you should contemplate that users could pass a vector of length > 1 (never underestimate end-users). If this happens, an error is given because `if()` can only accept 1-length vectors. You should do something like `if(!all(nuts_code %in% colnames(data)))`.

The “nuts.verbose” options should be mentioned in the help of each function. Otherwise, users can’t know what options they have about verbosity.

In the table about converting relative values, some `*` are interpreted by markdown as italics. This makes impossible to interpret the weighted average formulas. You should fix it, possibly converting to equations or to code. You will have to try.

#### Imports

All the scoped verbs (`*_at`) were superseeded in current versions of dplyr. Replace with new forms such as `summarize(across())`.

The pipe from `magrittr` and `as_tibble` from `tibble` package can be imported from `dplyr` and you can remove two packages from imports. You will depend on tibble and magrittr anyway, but not directly but trough `dplyr`.

Likewise, if you don’t want to depend on `stringr`, you can easily replace `str_remove` with `gsub()`, passing `""` as replacement; and `str_detect` with `grepl`.

#### `classify_nuts` (suggested name: `nuts_classify`)

-   L310, you should add the class “nuts.classified” instead of overwriting the class “list”: `class(x) <- c("new_class", class(x)`. Otherwise, if the object is no longer a list and only your custom class, this could break some methods for lists and raise unexpected behavior if the output of `classify_nuts` is used by something else than feed your other functions.

#### `convert_nuts_version` (suggested name: `nuts_convert_version`)

-   L72: There is `inherits(object, class)` to check a class. It’s more robust than `class(x) == "your_class"`.

-   L183: You could do the subset just after the `get()`. This way, you save loading the entire table in memory: `cross_walks <- get("cross_walks")[cross_walks$to_version == to_version, ]`.

#### `convert_nuts_level` (suggested name: `nuts_aggregate`)

-   In the vignettes, you explain that only aggregation is possible, which makes sense. However, this is not reflected in the help page. For instance, why `to_level` can be set to 3 if it’s not possible convert from level 1 or 2 to 3? You should add this in the description. My suggestion of name also makes this clearer.

L151: This chunk of code is repeated in `convert_nuts_version`. This is complex enough to avoid repeating yourself. I suggest moving all this chunk to a new function and call this function in both. This makes you code more maintainable.

Looking forward your responses!
