### 2024-02-19 #623 @maelle
Thanks a lot @jospueyo! :pray:

Could you please add the number of hours you spent reviewing near "Estimated hours spent reviewing:" in the template? Thanks!
