### 2024-02-19 70b2dec
Use underscores for variables in vignette
