+++ vignettes/nuts.Rmd
@@ -253,17 +253,17 @@ library(nuts)
 # Loading and subsetting Eurostat data
 data(patents, package = "nuts")
 
-pat.n2 <- patents %>% 
+pat_n2 <- patents %>% 
   filter(nchar(geo) == 4) # NUTS-2 values
 
-pat.n2.mhab.12.no <- pat.n2 %>%
+pat_n2_mhab_12_no <- pat_n2 %>%
   filter(unit == "P_MHAB") %>% # Patents per one million inhabitants
   filter(time == 2012) %>% # 2012
   filter(str_detect(geo, "^NO")) %>%  # Norway
   select(-unit)
 
 # Classifying the Data
-pat.classified <- classify_nuts(data = pat.n2.mhab.12.no,
+pat_classified <- classify_nuts(data = pat_n2_mhab_12_no,
                                 nuts_code = "geo")
 ```
 
@@ -274,20 +274,20 @@ The function returns a list with three items:
 Below, you see that all data entries correspond to the 2016 NUTS version.
 
 ```{r}
-pat.classified[["data"]]
+pat_classified[["data"]]
 ```
 
 2. The second item provides an overview of the share of matching NUTS codes for each of the five existing NUTS versions. The **overlap** is computed within country and possibly additional groups (if provided via the `group_vars` argument).  
 
 ```{r}
-pat.classified[["versions_data"]]
+pat_classified[["versions_data"]]
 ```
 
 3. The third item gives all NUTS codes that are **missing** across groups. Such missing codes might lead to conversion errors and are, by default, omitted from all conversion procedures. In our example, no NUTS codes are missing. 
 <!-- We recommend to check whether missing values for these NUTS codes can be replaced, perhaps with 0.  -->
 
 ```{r}
-pat.classified[["missing_data"]]
+pat_classified[["missing_data"]]
 ```
 
 
@@ -298,8 +298,8 @@ We provide the classified NUTS data, specify the target NUTS version for data tr
 
 ```{r}
 # Converting Data to 2021 NUTS version
-pat.converted <- convert_nuts_version(
-  data = pat.classified,
+pat_converted <- convert_nuts_version(
+  data = pat_classified,
   to_version = "2021",
   variables = c("values" = "relative")
 )
@@ -311,7 +311,7 @@ library(terra)
 no_2006 <-
   read_sf("NUTS_RG_20M_2016_3857.shp/NUTS_RG_20M_2016_3857.shp") %>%
   filter(LEVL_CODE == 2 & CNTR_CODE == "NO") %>%
-  full_join(pat.n2.mhab.12.no , by = c("NUTS_ID" = "geo"))
+  full_join(pat_n2_mhab_12_no , by = c("NUTS_ID" = "geo"))
 
 no_changes <- cross_walks %>%
   filter(
@@ -345,7 +345,7 @@ gg_2006 = ggplot() +
 no_2021 <-
   read_sf("NUTS_RG_20M_2021_3857.shp/NUTS_RG_20M_2021_3857.shp") %>%
   filter(LEVL_CODE == 2 & CNTR_CODE == "NO") %>%
-  full_join(pat.converted , by = c("NUTS_ID" = "to_code")) %>%
+  full_join(pat_converted , by = c("NUTS_ID" = "to_code")) %>%
   filter(NUTS_ID != "NO0B")
 
 gg_2021 = ggplot() +
@@ -375,11 +375,11 @@ ggarrange( gg_2006 , gg_2021 , nrow = 1 , common.legend = T, legend = "bottom")
[…]
