### 2024-02-19 a6d88cd
use !(inherits(object, class)) instead of class(x) == FALSE
