### 2024-02-19 abf45e4
Drop `== FALSE` conditions and replace with (`!`)
