### 2024-02-19 af1a529
Fix to `!` conditions
