### 2024-02-19 3a87491
avoid loading entire crosswalks table
