### 2024-02-20 38f8e17
introduce new function to test for multiple versions within group and select most frequent one or break
