+++ tests/testthat/test-nuts_convert_version.R
@@ -2,7 +2,7 @@
 #---------------------
 test_that("data input not valid", {
   expect_error(
-    convert_nuts_version(
+    nuts_convert_version(
       data = 1,
       to_version = "2021",
       variables = c("values" = "absolute")
@@ -14,8 +14,8 @@ test_that("data input not valid", {
 test_that("variables missing", {
   expect_error(
     manure_2_indic_DE_2003() %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(data = .,
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(data = .,
                            to_version = "2021")
   )
 })
@@ -23,8 +23,8 @@ test_that("variables missing", {
 test_that("variable not found", {
   expect_error(
     manure_2_indic_DE_2003() %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
         data = .,
         to_version = "2021",
         variables = c("valuess" = "absolute")
@@ -36,8 +36,8 @@ test_that("variable not found", {
 test_that("variable type not found", {
   expect_error(
     manure_2_indic_DE_2003() %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
         data = .,
         to_version = "2021",
         variables = c("values" = "absolutee")
@@ -49,8 +49,8 @@ test_that("variable type not found", {
 test_that("to_version invalid", {
   expect_error(
     manure_2_indic_DE_2003() %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
         data = .,
         to_version = "2020",
         variables = c("values" = "absolute")
@@ -62,8 +62,8 @@ test_that("to_version invalid", {
 test_that("weight invalid", {
   expect_error(
     manure_2_indic_DE_2003() %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
         data = .,
         to_version = "2021",
         variables = c("values" = "absolute"),
@@ -84,8 +84,8 @@ test_that("Mixing NUTS codes of different levels", {
         filter(grepl("^DE", geo)) %>%
         filter(geo != "DE") %>%
         filter(time == 2003) %>%
-        classify_nuts(nuts_code = "geo") %>%
-        convert_nuts_version(
+        nuts_classify(nuts_code = "geo") %>%
+        nuts_convert_version(
           data = .,
           to_version = "2021",
           variables = c("values" = "absolute")
@@ -102,8 +102,8 @@ test_that("Converter output spits out correct names", {
   expect_equal(
     manure_2_indic_DE_2003() %>%
       filter(!grepl("ZZ", geo)) %>%
-      classify_nuts(nuts_code = "geo") %>%
-      convert_nuts_version(
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
         to_version = "2021",
         variables = c("values" = "absolute")
       ) %>%
@@ -119,10 +119,10 @@ test_that("Grouped output equal to non-grouped output", {
       filter(grepl("DE", geo)) %>%
       filter(!grepl("ZZ", geo)) %>%
       filter(time %in% c(2000, 2010)) %>%
-      classify_nuts(nuts_code = "geo",
+      nuts_classify(nuts_code = "geo",
                     data = .,
                     group_vars = "time") %>%
-      convert_nuts_version(
+      nuts_convert_version(
         data = .,
[…]
