+++ man/nuts_test_multiple_versions.Rd
@@ -1,0 +1,15 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nuts_test_multiple_versions.R
+\name{nuts_test_multiple_versions}
+\alias{nuts_test_multiple_versions}
+\title{Helper function to test for multiple versions}
+\usage{
+nuts_test_multiple_versions(group_vars, multiple_versions, data_versions, data)
+}
+\value{
+A tibble containing NUTS codes and a message with the results of the test.
+}
+\description{
+\code{nuts_test_multiple_versions} is called from either \code{nuts_convert_version} or \code{nuts_aggregate}
+to selects the most frequent version within groups or throw an error.
+}

[…]
