+++ R/nuts_convert_version.R
@@ -207,15 +207,15 @@ nuts_convert_version <-
     # - Convert absolute values
     abs_data <- data %>%
       select(-all_of(rel_vars)) %>%
-      group_by_at(vars(any_of(c(
+      group_by(pick(c(
         "from_code", group_vars
-      )))) %>%
+      ))) %>%
       mutate(w = .data$w / sum(.data$w)) %>%
       ungroup() %>%
-      group_by_at(vars(any_of(c(
+      group_by(pick(c(
         "to_code", group_vars
-      )))) %>%
-      summarise_at(vars(all_of(abs_vars)), list( ~ sum(. * .data$w, na.rm = missing_rm))) %>%
+      ))) %>%
+      summarise(across(abs_vars, ~sum(.x * .data$w, na.rm = missing_rm))) %>%
       ungroup() %>%
       # - Add version
       mutate(to_version = to_version) %>%
@@ -224,11 +224,11 @@ nuts_convert_version <-
     # - Convert relative values
     rel_data <- data %>%
       select(-all_of(abs_vars)) %>%
-      group_by_at(vars(any_of(c(
+      group_by(pick(c(
         "to_code", group_vars
-      )))) %>%
-      summarise_at(vars(all_of(rel_vars)), list( ~ sum(. * .data$w, na.rm = missing_rm) /
-                                                   sum(.data$w))) %>%
+      ))) %>%
+      summarise(across(rel_vars, ~{sum(.x * .data$w, na.rm = missing_rm) /
+                                                   sum(.data$w)})) %>%
       ungroup() %>%
       # - Add version
       mutate(to_version = to_version) %>%
