### 2024-02-20 5013478
specify all options for arguments, i.e. error/most_frequent and most_recent/oldest
