### 2024-02-20 94accd4
document verbosity control
