+++ vignettes/nuts.Rmd
@@ -120,10 +120,10 @@ manure_indic <- manure %>%
 
 class <- manure_indic %>%
   distinct(geo, .keep_all = T) %>%
-  classify_nuts(data = ., nuts_code = "geo")
+  nuts_classify(data = ., nuts_code = "geo")
 
 transf <- class %>% 
-  convert_nuts_version( to_version = "2010" , variables = c( 'values' = 'absolute' ) , weight = 'artif_surf12' )
+  nuts_convert_version( to_version = "2010" , variables = c( 'values' = 'absolute' ) , weight = 'artif_surf12' )
 
 small <- manure_indic %>% 
   filter( geo %in% c( 'DED1' , 'DED3' ) )
@@ -226,15 +226,15 @@ gridExtra::grid.arrange( p_initial, p_step , p_final , nrow = 1 )
 
 The package comes with three main functions:
 
--   `classify_nuts()` detects the NUTS version(s) and level(s) of a data set. Its output can be directly fed into the two other functions.
+-   `nuts_classify()` detects the NUTS version(s) and level(s) of a data set. Its output can be directly fed into the two other functions.
 
--   `convert_nuts_version()` converts your data to a desired NUTS version (2006, 2010, 2013, 2016, 2021). This transformation works in any direction.
+-   `nuts_convert_version()` converts your data to a desired NUTS version (2006, 2010, 2013, 2016, 2021). This transformation works in any direction.
 
--   `convert_nuts_level()` aggregates data to some upper-level NUTS code, i.e., it transforms NUTS-3 data to the NUTS-2 or NUTS-1 level (but not vice versa).
+-   `nuts_aggregate()` aggregates data to some upper-level NUTS code, i.e., it transforms NUTS-3 data to the NUTS-2 or NUTS-1 level (but not vice versa).
 
 ## Workflow
 
-The conversion can only be conducted after classifying the NUTS version(s) and level(s) of your data using the function `classify_nuts()`. This step ensures the validity and completeness of your NUTS codes before proceeding with the conversion.
+The conversion can only be conducted after classifying the NUTS version(s) and level(s) of your data using the function `nuts_classify()`. This step ensures the validity and completeness of your NUTS codes before proceeding with the conversion.
 
 ```{r , echo=FALSE, out.width='60%', fig.align="center", fig.cap="Sequential workflow to convert regional NUTS data", fig.alt ="Flow diagram that shows that conversion functions are run after classification."}
 knitr::include_graphics("flow.png")
@@ -244,7 +244,7 @@ knitr::include_graphics("flow.png")
 
 ## Identifying NUTS version and level
 
-The `classify_nuts()` function's main purpose is to find the most suitable NUTS **version** and to identify the **level** of the data set. Below, you see an example using patent application data (per one million inhabitants) for Norway in 2012 at the NUTS-2 level. This data is provided by EUROSTAT.
+The `nuts_classify()` function's main purpose is to find the most suitable NUTS **version** and to identify the **level** of the data set. Below, you see an example using patent application data (per one million inhabitants) for Norway in 2012 at the NUTS-2 level. This data is provided by EUROSTAT.
 
 ```{r}
 # Load package
@@ -263,7 +263,7 @@ pat_n2_mhab_12_no <- pat_n2 %>%
   select(-unit)
 
 # Classifying the Data
-pat_classified <- classify_nuts(data = pat_n2_mhab_12_no,
+pat_classified <- nuts_classify(data = pat_n2_mhab_12_no,
                                 nuts_code = "geo")
 ```
 
@@ -298,7 +298,7 @@ We provide the classified NUTS data, specify the target NUTS version for data tr
 
 ```{r}
 # Converting Data to 2021 NUTS version
-pat_converted <- convert_nuts_version(
[…]
