+++ R/nuts-package.R
@@ -41,6 +41,8 @@
 #' @importFrom dplyr join_by
 #' @importFrom dplyr desc
 #' @importFrom dplyr arrange_at
+#' @importFrom dplyr across
+#' @importFrom dplyr pick
 #' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 ## usethis namespace: end
