+++ NAMESPACE
@@ -7,6 +7,7 @@ export(nuts_test_multiple_versions)
 import(cli)
 import(rlang)
 importFrom(dplyr,"%>%")
+importFrom(dplyr,across)
 importFrom(dplyr,all_of)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,any_of)
@@ -31,6 +32,7 @@ importFrom(dplyr,mutate)
 importFrom(dplyr,mutate_at)
 importFrom(dplyr,n)
 importFrom(dplyr,n_distinct)
+importFrom(dplyr,pick)
 importFrom(dplyr,pull)
 importFrom(dplyr,relocate)
 importFrom(dplyr,rename)
