+++ R/nuts_classify.R
@@ -6,7 +6,7 @@
 #' NUTS codes must be of the same level and need to be unique, unless additional grouping variables are specified. No
 #' duplicate NUTS codes within groups allowed.
 #' @param nuts_code Variable name containing NUTS codes
-#' @param group_vars Variable name(s) for classification within groups. Always computes overlap within country. `NULL`by default.
+#' @param group_vars Variable name(s) for classification within groups. Always computes overlap within country. `NULL` by default.
 #' @param ties Picks `'most_recent'` or `'oldest'` version when overlap is identical across multiple NUTS versions. `'most_recent'`
 #' by default.
 #'
