### 2024-02-21 203c787
add params to nuts_test_multiple_versions
