+++ tests/testthat/test-nuts_get_data.R
@@ -1,0 +1,10 @@
+# Run error tests
+#---------------------
+test_that("data input not valid", {
+  expect_error(
+    nuts_get_data(
+      data = 1
+    ),
+    "Input `data` must be a nuts.classified-object, not a number."
+  )
+})
