+++ vignettes/nuts.Rmd
@@ -82,7 +82,7 @@ For example, the German district *Northern Saxony* (*Nordsachsen*) is located wi
 -   NUTS-3: Districts
     -   DED53: Northern Saxony
 
-Since administrative boundaries in Europe change for demographic, economic, political or others reasons, there are five different **versions** of the NUTS Nomenclature (2006, 2010, 2013, 2016, and 2021). The current version, effective from 1 January 2021, lists 104 regions at NUTS-1, 283 regions at NUTS-2, 1 345 regions at NUTS-3 level[^2].
+Since administrative boundaries in Europe change for demographic, economic, political or other reasons, there are five different **versions** of the NUTS Nomenclature (2006, 2010, 2013, 2016, and 2021). The current version, effective from 1 January 2021, lists 104 regions at NUTS-1, 283 regions at NUTS-2, 1 345 regions at NUTS-3 level[^2].
 
 [^1]: [European Interinstitutional Style Guide](https://publications.europa.eu/code/en/en-5000600.htm) In the case of Greece for instance this code was changed from GR to EL in 2011. 
 
@@ -120,10 +120,17 @@ manure_indic <- manure %>%
 
 class <- manure_indic %>%
   distinct(geo, .keep_all = T) %>%
-  nuts_classify(data = ., nuts_code = "geo")
+  nuts_classify(
+    data = ., 
+    nuts_code = "geo"
+    )
 
 transf <- class %>% 
-  nuts_convert_version( to_version = "2010" , variables = c( 'values' = 'absolute' ) , weight = 'artif_surf12' )
+  nuts_convert_version( 
+    to_version = "2010", 
+    variables = c('values'='absolute'), 
+    weight = 'artif_surf12' 
+    )
 
 small <- manure_indic %>% 
   filter( geo %in% c( 'DED1' , 'DED3' ) )
@@ -244,7 +251,7 @@ knitr::include_graphics("flow.png")
 
 ## Identifying NUTS version and level
 
-The `nuts_classify()` function's main purpose is to find the most suitable NUTS **version** and to identify the **level** of the data set. Below, you see an example using patent application data (per one million inhabitants) for Norway in 2012 at the NUTS-2 level. This data is provided by EUROSTAT.
+The `nuts_classify()` function's main purpose is to find the most suitable NUTS **version** and to identify the **level** of the data set. Below, you see an example using patent application data (per one million inhabitants) for Norway in 2012 at the NUTS-2 level. This data is again provided by EUROSTAT.
 
 ```{r}
 # Load package
@@ -263,37 +270,42 @@ pat_n2_mhab_12_no <- pat_n2 %>%
   select(-unit)
 
 # Classifying the Data
-pat_classified <- nuts_classify(data = pat_n2_mhab_12_no,
-                                nuts_code = "geo")
+pat_classified <- nuts_classify(
+  data = pat_n2_mhab_12_no,
+  nuts_code = "geo"
+  )
 ```
 
-The function returns a list with three items:
+The function returns a list with three items. These items can be called directly from the output object (`data$...`) or retrieved using the three helper functions `nuts_get_data()`, `nuts_get_version()`, and `nuts_get_missing()`.
 
 1. The first item gives the **original data set** augmented with the columns `from_version`, `from_level`, and `country`, indicating the NUTS version that best suits the data. All functions of the package always group NUTS codes across **country names** which are automatically generated from the provided NUTS codes.
 
 Below, you see that all data entries correspond to the 2016 NUTS version.
 
 ```{r}
-pat_classified[["data"]]
+# pat_classified$data # Call list item directly or...
+nuts_get_data(pat_classified) # ...use helper function
 ```
 
[…]
