### 2024-02-23 2b5bb31
delete cli divs
