### 2024-02-23 b23507b
positive tests nuts_get_version()
