### 2024-02-23 #623 @nolwenn
## nuts package review

**Briefly describe any working relationship you have (had) with the package authors.**

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

\#### Functionality

The package could be really useful for many applications where spatial-temporal question is involved. At the time of my review, the authors have already considered some of the comments of reviewer 1 so you will excuse me if some of my comments might be outdated.

#### Installation instructions github

you have a error, an extra "/" . It should be `r pak::pak("AAoritz/nuts")` and not `r pak::pak("AAoritz/nuts/")`

At first load I had the message:

Error: package or namespace load failed for ‘nuts’ in loadNamespace(j \<- i[[1L]], c(lib.loc, .libPaths()), versionCheck = vI[[j]]): namespace ‘vctrs’ 0.6.3 is already loaded, but \>= 0.6.4 is required

I had to close and reopen R/Rstudio then it worked fine.

#### Vignette(s)

The Rmd file has been updated with the previous recommendations by @jospueyo concerning the naming of the function (which is good) but not the HTML version <https://aaoritz.github.io/nuts/articles/nuts-vignette.html>. The vignette in HTML version need to be rebuilt and push for online display.

In the vignette, the author did not explicitly call the libraries required for running the data management needed for the examples like the meta-library _tidyverse_ with _dplyr_ for the _filter()_ functions and other data management functions.

In the vignette, maps are displayed with various NUTS over time. Even so it is not the goal of the nuts package, a note mentioning which package(s) (or software) was(were) used to create those maps could be of interest for the reader.

The function nuts_test_multiple_versions do not have an examples. Would it be possible to have at least one?

After a nuts_convert_version(), it might be interesting to have a helping function that computes the estimated variations of the values between the different NUTS versions. It could be the difference (increase or decrease) in absolute or relative frequency at the higher NUT level. Like to get the 7.7% that you estimated in your first manure example (Spatial interpolation in a nutshell)

Hope it helps,

Best,
Nolwenn
