### 2024-02-23 #623 @krausewe
Amazing, @nolwenn. Thank you for the feedback!

(And sorry for already starting to revise the package while you were still reviewing.)
