### 2024-02-23 ebbb2b8
clean library calls in vignette
