### 2024-02-28 8a1d3dd
Provide short description how plots in vignette were made
