### 2024-02-28 764b0f4
fresh pkgdown, further links in readme, brackets in vignette
