+++ _pkgdown.yml
@@ -4,10 +4,14 @@ template:
 home:
   title: Convert European Regional Data in R
 reference:
-- title: Functions
+- title: Workhorse Functions
   desc:  Functions that classify and convert regional data
 - contents:
-    starts_with("nuts_classify|nuts_aggregate|nuts_convert|nuts_test_multiple_versions|nuts_get_data|nuts_get_version|nuts_get_missing")
+    starts_with("nuts_classify|nuts_aggregate|nuts_convert")
+- title: Additional Functions
+  desc:  Helper functions to classify and convert regional data
+- contents:
+    starts_with("nuts_test_multiple_versions|nuts_get_data|nuts_get_version|nuts_get_missing")
 - title: Data
   desc:  Data sets required for conversion and examples
 - contents:
