### 2024-02-28 813ad7c
add example for nuts_test_multiple_versions(), fix n rows dropped
