+++ R/nuts_test_multiple_versions.R
@@ -10,6 +10,21 @@
 #'
 #' @return A tibble containing NUTS codes and a message with the results of the test.
 #'
+#' @examples
+#' library(dplyr)
+#' df <- manure %>%
+#' filter(nchar(geo) == 5) %>%
+#' select(geo, indic_ag, values) %>%
+#' distinct(geo,  .keep_all = T) %>%
+#' nuts_classify(nuts_code = "geo",
+#'               group_vars = "indic_ag",
+#'               data = .)
+#'
+#' nuts_test_multiple_versions(group_vars = "indic_ag",
+#'                             multiple_versions = "most_frequent",
+#'                             data_versions = df$versions_data,
+#'                             data = df$data)
+#'
 #' @export
 
 nuts_test_multiple_versions = function(group_vars, multiple_versions, data_versions, data) {
@@ -56,8 +71,8 @@ nuts_test_multiple_versions = function(group_vars, multiple_versions, data_versi
       c("v" =  "{.blue Version is {.red unique}.}")
   }
 
-  data_list <- list(data, message_multiple_versions)
-  names(data_list) <- c("data", "message_multiple_versions")
+  data_list <- list(data, n_rows_dropped, message_multiple_versions)
+  names(data_list) <- c("data", "n_rows_dropped", "message_multiple_versions")
 
   return(data_list)
 }
