### 2024-03-01 330f5ca
fix nr of rows dropped in nuts_test_multiple_versions()
