### 2024-03-02 f052aa4
modify error message when user supplies multiple fields in single field arguments
