+++ tests/testthat/test-nuts_convert_version.R
@@ -55,7 +55,20 @@ test_that("to_version invalid", {
         to_version = "2020",
         variables = c("values" = "absolute")
       ),
-    "Input `to_version` invalid. Make sure it is a string and one of the version years 2006, 2010, 2013, 2016 or 2021."
+    "Input `to_version` invalid. Make sure it is a string and only one of the version years 2006, 2010, 2013, 2016 or 2021."
+  )
+})
+
+test_that("multiple to_versions provided", {
+  expect_error(
+    manure_2_indic_DE_2003() %>%
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
+        data = .,
+        to_version = c("2006", "2020"),
+        variables = c("values" = "absolute")
+      ),
+    "Input `to_version` invalid. Make sure it is a string and only one of the version years 2006, 2010, 2013, 2016 or 2021."
   )
 })
 
@@ -69,10 +82,24 @@ test_that("weight invalid", {
         variables = c("values" = "absolute"),
         weight = "pop19"
       ),
-    "Input `weight` invalid. Must be 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+    "Input `weight` invalid. Must be either 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
   )
 })
 
+
+test_that("multiple weights provided", {
+  expect_error(
+    manure_2_indic_DE_2003() %>%
+      nuts_classify(nuts_code = "geo") %>%
+      nuts_convert_version(
+        data = .,
+        to_version = "2021",
+        variables = c("values" = "absolute"),
+        weight = c("pop19", "areaKm")
+      ),
+    "Input `weight` invalid. Must be either 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+  )
+})
 
 
 test_that("Mixing NUTS codes of different levels", {
