### 2024-03-05 #623 @jospueyo
## Reviewer Response

This has been awesome. I enjoyed reviewing your package and monitoring its evolution. I hope you are truly happy with my suggestions, I am with your responses.

Regarding `.by` argument, I didn't know it was still experimental. Otherwise, I wouldn't recommend the change.

I also agree with your opinion about imputation of missing values.

#### Final approval (post-review)

Estimated hours spent reviewing: 3.5
