### 2024-03-11 #623 @nolwenn
## Reviewer Response

I enjoyed reviewing your package and was happy you found my comments valuable.  Sorry for the late replies to your answers.

About the remaining points of discussion:
1.  I have no longer the namespace error message. All seems working out well.

2. `r nuts_convert_version()`
- including an option in `r nuts_convert_version()` that could return e.g. a column in the converted dataset containing more details on the conversion could indeed be an interesting way of doing implementing it.
- an example in the vignette on how to  extracted, by sub-setting and transforming the tibble data(cross_walks) is also nice

3. Missing regions : weights missing due to missing regions by changing the `r missing_rm` option to a threshold at which the users feel confident to assume 0 for the missing regions is a good idea to implement.

#### Final approval (post-review)

- [ x]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 1 hour
