### 2024-03-12 5af7b8f
share of missing weights option to `nuts_convert` and `nuts_aggregate`, elaborate internatl data in vignette
