### 2024-03-12 89c84e3
elaborate in vignette how internal conversion tables can be accessed and used
