### 2024-03-12 #623 @AAoritz
One question on verbosity control @maelle. We thought that our current implementation using `rlang::local_options()` based on your recommendations would set verbosity as a default. However, verbosity has to be switched on to function. Do you understand why?
