### 2024-03-14 #623 @maelle
In
https://github.com/AAoritz/nuts/blob/cf42cce8e522d4bd9ae1e8506024620dd964cebe/R/nuts_convert_version.R#L299

```r
is_verbose_mode <- (getOption("nuts.verbose", "quiet") == "verbose")
```

The default is "quiet", not "verbose".
