### 2024-03-14 #623 @AAoritz
Makes sense... we set the default to `verbose`. Thank you @maelle for this, your super helpful comments and your work as the editor during the review process! @krausewe and I learned a lot!
