### 2024-03-14 #623 @jospueyo
> > @nolwenn and @jospueyo, many thanks again! Please let us know if we can add you in the meta description as reviewers of the package. @maelle can we include you as well?!
>
> Thanks for proposing. I do not know the convention but do as you like. As @maelle I am already happy to know that I helped.

I'd be honoured to be recognized for my tiny contribution to your so useful package.
